#269 Made changes as per the review
