+++ src/main/java/edu/ie3/datamodel/models/result/system/ChpResult.java
@@ -13,8 +13,6 @@ import tech.units.indriya.ComparableQuantity;
 /** Represents calculation results of a {@link edu.ie3.datamodel.models.input.system.ChpInput} */
 public class ChpResult extends SystemParticipantWithHeatResult {
 
-  private final ComparableQuantity<Power> qDot;
-
   /**
    * Standard constructor with automatic uuid generation.
    *
@@ -31,7 +29,6 @@ public class ChpResult extends SystemParticipantWithHeatResult {
       ComparableQuantity<Power> q,
       ComparableQuantity<Power> qDot) {
     super(time, inputModel, p, q, qDot);
-    this.qDot = qDot;
   }
 
   /**
@@ -53,7 +50,6 @@ public class ChpResult extends SystemParticipantWithHeatResult {
       ComparableQuantity<Power> q,
       ComparableQuantity<Power> qDot) {
     super(uuid, time, inputModel, p, q, qDot);
-    this.qDot = qDot;
   }
 
   @Override
@@ -70,7 +66,7 @@ public class ChpResult extends SystemParticipantWithHeatResult {
         + ", q="
         + getQ()
         + ", qDot="
-        + qDot
+        + getqDot()
         + '}';
   }
 }
