+++ src/main/java/edu/ie3/datamodel/models/result/system/HpResult.java
@@ -6,16 +6,12 @@
 package edu.ie3.datamodel.models.result.system;
 
 import java.time.ZonedDateTime;
-import java.util.Objects;
 import java.util.UUID;
-import javax.measure.Quantity;
 import javax.measure.quantity.Power;
 import tech.units.indriya.ComparableQuantity;
 
 /** Represents calculation results of a {@link edu.ie3.datamodel.models.input.system.HpInput} */
-public class HpResult extends SystemParticipantResult {
-
-  private final Quantity<Power> qDot;
+public class HpResult extends SystemParticipantWithHeatResult {
 
   /**
    * Standard constructor with automatic uuid generation.
@@ -31,9 +27,8 @@ public class HpResult extends SystemParticipantResult {
       UUID inputModel,
       ComparableQuantity<Power> p,
       ComparableQuantity<Power> q,
-      Quantity<Power> qDot) {
-    super(time, inputModel, p, q);
-    this.qDot = qDot;
+      ComparableQuantity<Power> qDot) {
+    super(time, inputModel, p, q, qDot);
   }
 
   /**
@@ -53,27 +48,8 @@ public class HpResult extends SystemParticipantResult {
       UUID inputModel,
       ComparableQuantity<Power> p,
       ComparableQuantity<Power> q,
-      Quantity<Power> qDot) {
-    super(uuid, time, inputModel, p, q);
-    this.qDot = qDot;
-  }
-
-  public Quantity<Power> getqDot() {
-    return qDot;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-    if (!super.equals(o)) return false;
-    HpResult hpResult = (HpResult) o;
-    return qDot.equals(hpResult.qDot);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(super.hashCode(), qDot);
+      ComparableQuantity<Power> qDot) {
+    super(uuid, time, inputModel, p, q, qDot);
   }
 
   @Override
@@ -90,7 +66,7 @@ public class HpResult extends SystemParticipantResult {
         + ", q="
         + getQ()
         + ", qDot="
-        + qDot
+        + getqDot()
         + '}';
   }
 }
